Update to Java 8 to enable TLS 1.2
kb-sdk test fails in both appdev and CI with the error
```
Object #1, Ecoli_042.genome failed type checking:
    [junit] instance type (string) does not match any allowed primitive type (allowed: ["array"]), at /features/0/aliases/0
```
so I assume the tests are just broken.

When run as a job in the narrative it fails with:
```
java.lang.NullPointerException
	at us.kbase.common.service.JsonServerServlet.validateToken(JsonServerServlet.java:702)
	at us.kbase.common.service.JsonServerServlet.processRpcCall(JsonServerServlet.java:563)
	at us.kbase.common.service.JsonServerServlet.processRpcCall(JsonServerServlet.java:499)
	at genomeproteomecomparison.GenomeProteomeComparisonServer.main(GenomeProteomeComparisonServer.java:70)
```
So presumably the type checking error is masking the token validation error.

Given all that, don't worry about the tests and test in the narrative.